// confmod-app/frontend/src/app/openapi/model/createModelAndConfig.ts
/**
 * Confmod Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RawConfmodConfigInput } from './rawConfmodConfigInput';
import { RawConfmodModelInput } from './rawConfmodModelInput';


export interface CreateModelAndConfig { 
    name: string;
    model?: RawConfmodModelInput | null;
    config?: RawConfmodConfigInput | null;
}