// confmod-app/frontend/src/app/openapi/model/rawConfmodModelOutput.ts
/**
 * Confmod Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RawConfmodHeaderOutput } from './rawConfmodHeaderOutput';
import { RawConfmodModelInputPayloadInnerInner } from './rawConfmodModelInputPayloadInnerInner';


export interface RawConfmodModelOutput { 
    head: RawConfmodHeaderOutput;
    payload: Array<Array<RawConfmodModelInputPayloadInnerInner>>;
}